Fix: Avoid recursive calling of MaxLinkTime() by adding a maxlinktime-cache
When having a project with many imported files which import each other too this leads to an exorbitant call to MaxLinkTime() as it calls it again and again for all used files (even if already calculated before). In my game "TVTower" this lead to 2456886754 (!) calls while the cache one only needs 886 and 1419 cache calls.

Compilation ("STAGE_LINK") time is reduced from ~157s to ~0.6s.